+++ CHANGELOG.md
@@ -5,7 +5,8 @@ Version 2.0.1:
 * Added a `CanReleaseResource` object for `scala.io.Source`, which doesn't
   extend `java.io.Closeable`. This allows `withResource()` to work with
   `scala.io.Source` objects, without the user having to write a custom
-  `CanReleaseResource[Source]` object.
+  `CanReleaseResource[Source]` object. (Pull request from
+  [Adam Lesperance](https://github.com/lespea).)
 
 Version 2.0.0 (Incompatible API changes):